requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
autopep8==2.0.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
distlib==0.3.7
docopt==0.6.2
fastapi==0.100.0
filelock==3.12.2
flake8==6.1.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
identify==2.5.26
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pipreqs==0.4.13
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
psycopg2==2.9.6
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pyflakes==3.1.0
pytest==7.4.0
python-dotenv==1.0.0
pyupgrade==3.10.1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tokenize-rt==5.2.0
types-cachetools==5.3.0.6
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.3
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
virtualenv==20.24.2
yarg==0.1.9
